# Modify benchmark to load sources only once