Long time ago, after a regrettable update, the benchmark module ended up reloading sources for every output line, i.e. every compression level.

While the loading delay itself is likely tolerable,
the main issue is that the `--quiet` mode now also displays a loading summary between each compression line. 
This wasn't the original intention, which is to produce a compact view of all compression results.

This is fixed in this version,
where sources are loaded only once, for all compression levels, and loading summary is now only displayed once, as intended.

Before (`dev`) : 
```
zstd -b1e5 calgary.tar enwik7 -q -S
bench 1.5.6 : input 3153920 bytes, 0 seconds, 0 KB blocks
-1      1138663 (2.770)  43.59 MB/s  190.2 MB/s  calgary.tar
bench 1.5.6 : input 3153920 bytes, 0 seconds, 0 KB blocks
-2      1063638 (2.965)  42.62 MB/s  137.6 MB/s  calgary.tar
bench 1.5.6 : input 3153920 bytes, 0 seconds, 0 KB blocks
-3      1020606 (3.090)  33.07 MB/s  125.3 MB/s  calgary.tar
bench 1.5.6 : input 3153920 bytes, 0 seconds, 0 KB blocks
-4      1011132 (3.119)  32.72 MB/s  121.2 MB/s  calgary.tar
bench 1.5.6 : input 3153920 bytes, 0 seconds, 0 KB blocks
-5       984050 (3.205)  16.20 MB/s  119.4 MB/s  calgary.tar
bench 1.5.6 : input 10000000 bytes, 0 seconds, 0 KB blocks
-1      4105389 (2.436)  37.88 MB/s  168.3 MB/s  enwik7
bench 1.5.6 : input 10000000 bytes, 0 seconds, 0 KB blocks
-2      3758552 (2.661)  38.16 MB/s  117.6 MB/s  enwik7
bench 1.5.6 : input 10000000 bytes, 0 seconds, 0 KB blocks
-3      3565658 (2.805)  30.16 MB/s  105.9 MB/s  enwik7
bench 1.5.6 : input 10000000 bytes, 0 seconds, 0 KB blocks
-4      3493017 (2.863)  30.88 MB/s  102.3 MB/s  enwik7
bench 1.5.6 : input 10000000 bytes, 0 seconds, 0 KB blocks
-5      3392330 (2.948)  15.05 MB/s   99.5 MB/s  enwik7
```

fixed in this PR : 
```
zstd -b1e5 calgary.tar enwik7 -q -S
bench 1.5.6 : input 3153920 bytes, 0 seconds, 0 KB blocks          
-1      1138663 (2.770) 316.74 MB/s 1114.0 MB/s  calgary.tar                                                                          
-2      1063638 (2.965) 222.79 MB/s  951.7 MB/s  calgary.tar                                                                          
-3      1020606 (3.090) 165.58 MB/s  909.1 MB/s  calgary.tar       
-4      1011132 (3.119) 156.22 MB/s  894.6 MB/s  calgary.tar       
-5       984050 (3.205)  92.40 MB/s  868.1 MB/s  calgary.tar
bench 1.5.6 : input 10000000 bytes, 0 seconds, 0 KB blocks      
-1      4105389 (2.436) 276.81 MB/s 1015.4 MB/s  enwik7
-2      3758552 (2.661) 194.60 MB/s  864.6 MB/s  enwik7
-3      3565658 (2.805) 149.71 MB/s  803.3 MB/s  enwik7
-4      3493017 (2.863) 145.57 MB/s  807.5 MB/s  enwik7
-5      3392330 (2.948)  87.99 MB/s  788.0 MB/s  enwik7

```